+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
-## WindMenu (v.0.1.1)
+## WindMenu (v.0.2.0)
 
+- Format KeyCombo commands as `name [keys]`
 - Small selection bug fixed